Extract request logger to own module
